requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
pydantic==1.10.13
python-multipart==0.0.6
pdfplumber==0.10.3
tabula-py==2.9.0
pandas==2.0.3
python-dotenv==1.0.0
requests==2.31.0
numpy==1.24.3
jinja2==3.1.2
aiofiles==23.2.1